Add character encoding declaration
Technically, a character encoding declaration isn’t required. The HTML
Standard[1] says:
	“If an HTML document does not start with a BOM, and its
	encoding is not explicitly given by Content-Type metadata, and
	the document is not an iframe srcdoc document, then the encoding
	must be specified using a meta element with a charset attribute or a
	meta element with an http-equiv attribute in the Encoding
	declaration state.”[2]

index.html is an HTML document (or at least it should be considered one
in order to be interpreted properly)[3]. Before this change,
	- It didn’t start with a BOM
	- Its encoding was explicitly given by Content-Type metadata, but
	  this only happens after it gets uploaded to GitHub Pages. When
	  testing the page out offline, there is might not be
	  Content-Type metadata. Plus, the page might get moved to a
	  different server or protocol that doesn’t provide such
	  metadata.
	- It wasn’t an iframe srcdoc document.

This suggests that the document needs to have a character encoding
declaration, but only in some situations. In my opinion, it’s better to
just include a character encoding declaration because you can’t
guarantee that the Content-Type metadata will be there or accurate.

The HTML Standard also says
	“The element containing the character encoding declaration must be
	serialized completely within the first 1024 bytes of the document.”[2]

The text of the MIT license is more than 1024 bytes long. This means
that either
	A. The license must be placed after an encoding declaration, or
	B. A BOM must be used.

I decided to use a BOM because, traditionally, licensing information is
put right at the top of source files.

Links:
1. <https://html.spec.whatwg.org/multipage/>
2. <https://html.spec.whatwg.org/multipage/semantics.html#charset>
3. <https://dom.spec.whatwg.org/#html-document>